Move things from target.cpp to dmesg.cpp
